requirements.txt
# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.1.0

# Production server
gunicorn==21.2.0

# PostgreSQL
psycopg2-binary==2.9.9

# Static files serving
whitenoise==6.6.0

# Environment variables
python-decouple==3.8
dj-database-url==2.1.0
